+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Changelog
 
-## v1.0.5 - 2023-12-09
+## v1.0.5 - 2023-12-10
 
 ### Added
 
 - Added an icon to the AddOn List
-- Added support for Mythic0 and Heroic vault values/tooltips
+- Added support for Heroic and Mythic dungeon runs
 
 ### Updated
 
@@ -16,6 +16,7 @@
 - Fixed a bug showing incorrect kills for the Council and Larodar encounters
 - Rating tooltip no longer shows +0 as highest key level
 - The vault tooltip no longer shows improvement text if all weekly dungeon runs are level 20+
+- The "Rewards" label now hides properly when looting the vault
 
 ## v1.0.4 - 2023-11-22